// apps/bycut/src/types/timeline.ts
export interface VideoElement extends BaseTimelineElement {
  type: 'video'
  mediaId: string
  muted?: boolean
  hidden?: boolean
  transform: Transform
  opacity: number
  playbackRate?: number
  reversed?: boolean
}

export interface ImageElement extends BaseTimelineElement {
  type: 'image'
  mediaId: string
  hidden?: boolean
  transform: Transform
  opacity: number
}

export interface TextStroke {
  color: string
  width: number
}

export interface TextShadow {
  color: string
  offsetX: number
  offsetY: number
  blur: number
}

export interface TextElement extends BaseTimelineElement {
  type: 'text'
  content: string
  fontSize: number
  fontFamily: string
  color: string
  backgroundColor: string
  textAlign: 'left' | 'center' | 'right'
  fontWeight: 'normal' | 'bold'
  fontStyle: 'normal' | 'italic'
  textDecoration: 'none' | 'underline' | 'line-through'
  hidden?: boolean
  transform: Transform
  opacity: number
  stroke?: TextStroke
  shadow?: TextShadow
  boxWidth?: number
  backgroundBorderRadius?: number
  backgroundOpacity?: number
  backgroundPaddingX?: number
  backgroundPaddingY?: number
}

export interface StickerElement extends BaseTimelineElement {
  type: 'sticker'
  iconName: string
  hidden?: boolean
  transform: Transform
  opacity: number
  color?: string
}

export type TimelineElement =
  | AudioElement
  | VideoElement
  | ImageElement
  | TextElement
  | StickerElement

export type ElementType = TimelineElement['type']

export type CreateUploadAudioElement = Omit<UploadAudioElement, 'id'>
export type CreateLibraryAudioElement = Omit<LibraryAudioElement, 'id'>
export type CreateAudioElement =
  | CreateUploadAudioElement
  | CreateLibraryAudioElement
export type CreateVideoElement = Omit<VideoElement, 'id'>
export type CreateImageElement = Omit<ImageElement, 'id'>
export type CreateTextElement = Omit<TextElement, 'id'>
export type CreateStickerElement = Omit<StickerElement, 'id'>
export type CreateTimelineElement =
  | CreateAudioElement
  | CreateVideoElement
  | CreateImageElement
  | CreateTextElement
  | CreateStickerElement